# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS API Documentation",
        "description": "HRMS Backend API documentation with OpenAPI/Swagger",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/lookups": {
            "get": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Get all lookup values organized by category",
                "description": "Returns all system lookup values grouped by their respective categories",
                "operationId": "getLookups",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "gender": {
                                            "description": "List of gender options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "subsidiary": {
                                            "description": "List of subsidiary options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "employee_status": {
                                            "description": "List of employee status options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "nationality": {
                                            "description": "List of nationality options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "religion": {
                                            "description": "List of religion options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "marital_status": {
                                            "description": "List of marital status options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "site": {
                                            "description": "List of site options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "user_status": {
                                            "description": "List of user status options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "interview_mode": {
                                            "description": "List of interview mode options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "interview_status": {
                                            "description": "List of interview status options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "identification_types": {
                                            "description": "List of identification type options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        },
                                        "employment_type": {
                                            "description": "List of employment type options",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lookup"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Create a new lookup value",
                "description": "Stores a new lookup value in the system",
                "operationId": "storeLookup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Lookup type",
                                        "type": "string",
                                        "example": "gender"
                                    },
                                    "value": {
                                        "description": "Display value",
                                        "type": "string",
                                        "example": "Non-binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lookup created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lookup created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lookup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lookups/{id}": {
            "get": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Get a specific lookup value",
                "description": "Returns details for a specific lookup value",
                "operationId": "showLookup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lookup ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lookup"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lookup not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Update a lookup value",
                "description": "Updates an existing lookup value in the system",
                "operationId": "updateLookup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lookup ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "description": "Lookup type",
                                        "type": "string",
                                        "example": "gender"
                                    },
                                    "value": {
                                        "description": "Display value",
                                        "type": "string",
                                        "example": "Updated Value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lookup updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lookup updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lookup"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lookup not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Delete a lookup value",
                "description": "Removes a lookup value from the system",
                "operationId": "deleteLookup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lookup ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lookup deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lookup deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lookup not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lookups/type/{type}": {
            "get": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Get lookup values by type",
                "description": "Returns all lookup values for a specific type",
                "operationId": "getLookupsByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Lookup type (e.g., gender, nationality)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lookup"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No lookups found for this type"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lookup": {
                "title": "Lookup",
                "description": "Lookup model for various system reference data",
                "properties": {
                    "id": {
                        "description": "Lookup ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "description": "Type of lookup (gender, subsidiary, etc.)",
                        "type": "string"
                    },
                    "value": {
                        "description": "Display value",
                        "type": "string"
                    },
                    "created_by": {
                        "description": "User who created the record",
                        "type": "string",
                        "nullable": true
                    },
                    "updated_by": {
                        "description": "User who last updated the record",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}